# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "QAA'A API",
        "version": "1.0.0",
        "description": "API for managing hall bookings and reservations"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000",
            "description": "Local Development Server"
        }
    ],
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "token"
            }
        }
    },
    "paths": {
...
        "/api/halls": {
            "get": {
                "tags": ["Halls"],
                "summary": "Get all halls",
                "operationId": "getAllHalls",
                "responses": {
                    "200": {
                        "description": "List of all halls"
                    }
                }
            },
            "post": {
                "tags": ["Halls"],
                "summary": "Create new hall",
                "operationId": "createHall",
                "security": [
                    {
                        "sanctum": []
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "name": {
                                        "type": "string"
                                    },
                                    "location": {
                                        "type": "string"
                                    },
                                    "capacity": {
                                        "type": "integer"
                                    },
                                    "price_per_hour": {
                                        "type": "number"
                                    }
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Hall created successfully"
                    }
                }
            }
        },
...
    }
}